backend/server/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "3.13"
anyio==4.8.0 ; python_version >= "3.10" and python_version < "3.13"
attrs==25.1.0 ; python_version >= "3.10" and python_version < "3.13"
automat==24.8.1 ; python_version >= "3.10" and python_version < "3.13"
bcrypt==4.2.1 ; python_version >= "3.10" and python_version < "3.13"
bunnet==1.3.0 ; python_version >= "3.10" and python_version < "3.13"
certifi==2025.1.31 ; python_version >= "3.10" and python_version < "3.13"
cffi==1.17.1 ; python_version >= "3.10" and python_version < "3.13"
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_version < "3.13"
click==8.1.8 ; python_version >= "3.10" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.13" and (platform_system == "Windows" or sys_platform == "win32")
configparser==7.1.0 ; python_version >= "3.10" and python_version < "3.13"
constantly==23.10.4 ; python_version >= "3.10" and python_version < "3.13"
contourpy==1.3.1 ; python_version >= "3.10" and python_version < "3.13"
coverage==7.6.12 ; python_version >= "3.10" and python_version < "3.13"
croniter==2.0.7 ; python_version >= "3.10" and python_version < "3.13"
cryptography==44.0.1 ; python_version >= "3.10" and python_version < "3.13"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "3.13"
dnspython==2.7.0 ; python_version >= "3.10" and python_version < "3.13"
ecdsa==0.19.0 ; python_version >= "3.10" and python_version < "3.13"
email-validator==2.2.0 ; python_version >= "3.10" and python_version < "3.13"
exceptiongroup==1.2.2 ; python_version == "3.10"
fastapi-cli==0.0.7 ; python_version >= "3.10" and python_version < "3.13"
fastapi==0.111.1 ; python_version >= "3.10" and python_version < "3.13"
filelock==3.17.0 ; python_version >= "3.10" and python_version < "3.13"
fonttools==4.56.0 ; python_version >= "3.10" and python_version < "3.13"
future==1.0.0 ; python_version >= "3.10" and python_version < "3.13"
gunicorn==22.0.0 ; python_version >= "3.10" and python_version < "3.13"
h11==0.14.0 ; python_version >= "3.10" and python_version < "3.13"
httpcore==1.0.7 ; python_version >= "3.10" and python_version < "3.13"
httptools==0.6.4 ; python_version >= "3.10" and python_version < "3.13"
httpx==0.28.1 ; python_version >= "3.10" and python_version < "3.13"
hyperlink==21.0.0 ; python_version >= "3.10" and python_version < "3.13"
idna==3.10 ; python_version >= "3.10" and python_version < "3.13"
incremental==24.7.2 ; python_version >= "3.10" and python_version < "3.13"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "3.13"
jinja2==3.1.5 ; python_version >= "3.10" and python_version < "3.13"
kiwisolver==1.4.8 ; python_version >= "3.10" and python_version < "3.13"
lazy-model==0.2.0 ; python_version >= "3.10" and python_version < "3.13"
markdown-it-py==3.0.0 ; python_version >= "3.10" and python_version < "3.13"
markupsafe==3.0.2 ; python_version >= "3.10" and python_version < "3.13"
matplotlib==3.8.0 ; python_version >= "3.10" and python_version < "3.13"
mdurl==0.1.2 ; python_version >= "3.10" and python_version < "3.13"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "3.13"
mypy==1.15.0 ; python_version >= "3.10" and python_version < "3.13"
numpy==1.26.0 ; python_version >= "3.10" and python_version < "3.13"
packaging==23.2 ; python_version >= "3.10" and python_version < "3.13"
paramiko==3.5.1 ; python_version >= "3.10" and python_version < "3.13"
passlib==1.7.4 ; python_version >= "3.10" and python_version < "3.13"
pillow==11.1.0 ; python_version >= "3.10" and python_version < "3.13"
pluggy==1.5.0 ; python_version >= "3.10" and python_version < "3.13"
pyasn1-modules==0.4.1 ; python_version >= "3.10" and python_version < "3.13"
pyasn1==0.4.8 ; python_version >= "3.10" and python_version < "3.13"
pycparser==2.22 ; python_version >= "3.10" and python_version < "3.13"
pydantic-core==2.10.1 ; python_version >= "3.10" and python_version < "3.13"
pydantic==2.4.2 ; python_version >= "3.10" and python_version < "3.13"
pygments==2.19.1 ; python_version >= "3.10" and python_version < "3.13"
pylabrad==0.98.2 ; python_version >= "3.10" and python_version < "3.13"
pymongo==4.11.1 ; python_version >= "3.10" and python_version < "3.13"
pynacl==1.5.0 ; python_version >= "3.10" and python_version < "3.13"
pyopenssl==25.0.0 ; python_version >= "3.10" and python_version < "3.13"
pyparsing==3.2.1 ; python_version >= "3.10" and python_version < "3.13"
pytest-cov==5.0.0 ; python_version >= "3.10" and python_version < "3.13"
pytest-env==1.1.5 ; python_version >= "3.10" and python_version < "3.13"
pytest-mypy==0.10.3 ; python_version >= "3.10" and python_version < "3.13"
pytest==8.3.4 ; python_version >= "3.10" and python_version < "3.13"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "3.13"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "3.13"
python-jose==3.4.0 ; python_version >= "3.10" and python_version < "3.13"
python-multipart==0.0.20 ; python_version >= "3.10" and python_version < "3.13"
pytz==2025.1 ; python_version >= "3.10" and python_version < "3.13"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "3.13"
requests==2.32.3 ; python_version >= "3.10" and python_version < "3.13"
rich-toolkit==0.13.2 ; python_version >= "3.10" and python_version < "3.13"
rich==13.9.4 ; python_version >= "3.10" and python_version < "3.13"
rsa==4.9 ; python_version >= "3.10" and python_version < "3.13"
ruff==0.5.7 ; python_version >= "3.10" and python_version < "3.13"
service-identity==24.2.0 ; python_version >= "3.10" and python_version < "3.13"
setuptools-scm==8.2.0 ; python_version >= "3.10" and python_version < "3.13"
setuptools==75.8.2 ; python_version >= "3.10" and python_version < "3.13"
shellingham==1.5.4 ; python_version >= "3.10" and python_version < "3.13"
six==1.17.0 ; python_version >= "3.10" and python_version < "3.13"
slack-sdk==3.34.0 ; python_version >= "3.10" and python_version < "3.13"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "3.13"
starlette==0.37.2 ; python_version >= "3.10" and python_version < "3.13"
toml==0.10.2 ; python_version >= "3.10" and python_version < "3.13"
tomli==2.2.1 ; python_version == "3.10"
tqdm==4.67.1 ; python_version >= "3.10" and python_version < "3.13"
twisted==24.11.0 ; python_version >= "3.10" and python_version < "3.13"
typer==0.14.0 ; python_version >= "3.10" and python_version < "3.13"
types-paramiko==3.5.0.20240928 ; python_version >= "3.10" and python_version < "3.13"
types-python-dateutil==2.9.0.20241206 ; python_version >= "3.10" and python_version < "3.13"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "3.13"
urllib3==2.3.0 ; python_version >= "3.10" and python_version < "3.13"
uvicorn==0.30.6 ; python_version >= "3.10" and python_version < "3.13"
uvloop==0.21.0 ; python_version >= "3.10" and python_version < "3.13" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
watchfiles==1.0.4 ; python_version >= "3.10" and python_version < "3.13"
websockets==15.0 ; python_version >= "3.10" and python_version < "3.13"
zope-interface==7.2 ; python_version >= "3.10" and python_version < "3.13"
